requirements.txt
fastapi
uvicorn
httpx
python-multipart
Pillow
requests
python-dotenv
